put in place a "doOnce" directive that applies to one line only
something needs to be fixed in Chrome where the arguments of a function
are evaluated even if the function is not defined. This is a problem
because

doO alert 'miao'

is translated into

doO (alert ('miao'))

which alerts miao even if doO is not defined. This is a problem for how
the live coding might work.