the order of the css in the grunt targets does matter...
simpleModalBasic.css and simpleModalDemo.css must be last, probably
because something in them overrides the contents of the previous ones.
If left unspecified, in my machine the styles of the modal windows are
mangled, and the stat.js fps is taller than the menu bar. Not an
elegant solution, but an effective stop-gap one, is to give a specific
order.