Replacing the eval with function creation
Currently the patch code is wrapped in a function creator, compiled from
coffeescript, and then evaluated to create the draw function.

This change removes that coffeescript function wrapped and instead
passed the compiled javascript into a new Function object.

By assigning this new function object to the draw var the patch works
exactly as normal but in future it will be easier to control the scope
that the draw function is created and run in