Removing call to doOnceOccurrencesLineNumbers
doOnceOccurrencesLineNumbers is now an attribute of the CodeTransform
object so livecodelab.js calls it via there.

also added an addDoOnce function to CodeTransformer. This is added to
the global scope and is used to add the do once lines from evaled
scripts, instead of directly pushing them into the array