cleaned up the start sound. See extended description for details.
The second part of the sound had the original artifacts due to the
8-bit synth (fade-out would produce some audible clicks when sound is
brought to each lower volume level). Re-synthesized it with audacity -
it's a 10396Hz square wave with no antialiasing (which makes the wave
somewhat of a smoothed square wave). The applied a fade-out. Then also
applied a phaser over the entire sound to give it a bit more of a 3d
effect (since it's a stereo sound!)